+++ src/App.js
@@ -1,12 +1,10 @@
 import { useState } from 'react';
-
-import TypingField from './components/TypingField';
-import HomepageHeader from './components/HomepageHeader';
-import LayoutWrapper from './components/LayoutWrapper';
 import { Routes, Route } from 'react-router-dom';
-import About from './components/About/About';
-import Login from './components/Login';
+
 import { UserContext } from './helpers/context';
+import About from './routes/About';
+import Login from './routes/Login';
+import Home from './routes/Home';
 
 // parent component for all other components
 export default function App() {
@@ -17,25 +15,8 @@ export default function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="about" element={<About />} />
-        <Route
-          path="/login/*"
-          element={
-            <UserContext.Provider value={{ user, setUser }}>
-              <Login />
-            </UserContext.Provider>
-          }
-        />
+        <Route path="login" element={<Login />} />
       </Routes>
     </UserContext.Provider>
-  );
-}
-
-// !! needs refactoring to another file
-function Home() {
-  return (
-    <LayoutWrapper>
-      <HomepageHeader />
-      <TypingField />
-    </LayoutWrapper>
   );
 }
