+++ src/components/LayoutWrapper.jsx
@@ -1,37 +1,28 @@
 // wraps all components
 // imported into App.js
 import SectionContainer from './SectionContainer';
-import HomepageHeader from './HomepageHeader';
 
 const LayoutWrapper = ({ children }) => {
   return (
     <SectionContainer>
       <div className="flex h-screen flex-col justify-between font-sans">
-        <header className="flex items-center justify-between py-10">
+        <nav className="flex items-center justify-between py-10">
           <div>
             <div className="flex items-center justify-between">
               <div className="mr-3">CTHULHU TEACHES TYPES</div>
             </div>
           </div>
           <div className="flex items-center text-base leading-5">
-            <div className="hidden sm:block">
-              {['About'].map((link) => (
-                <a key={link} href={'/about'} className="p-1 font-medium sm:p-4">
-                  {link}
-                </a>
-              ))}
-            </div>
-            <div className="hidden sm:block">
-              {['Login'].map((link) => (
-                <a key={link} href={'/login'} className="p-1 font-medium sm:p-4">
-                  {link}
-                </a>
-              ))}
-            </div>
+            <a href={'/about'} className="p-1 font-medium sm:p-4">
+              About
+            </a>
+
+            <a href={'/login'} className="p-1 font-medium sm:p-4">
+              Login
+            </a>
           </div>
-        </header>
+        </nav>
         <main className="mb-auto">{children}</main>
-        {/* <Footer /> */}
       </div>
     </SectionContainer>
   );
